Move tests before refactoring them
